Complete basic functionality for meals
Meals as well as foods can be added without any issue onto the screen. There is a problem with updating the macro footer after adding foods, however. If separate meals are created, and a food is added to each meal, then the macro footer will update. If, however, two foods are added under a single meal, the macro footer ends up resetting to the value of the last food added to that meal.